14 16

+++ factory/ComplexAbstractFactory.java
@@ -2,5 +2,5 @@ package factory;
 
 public interface ComplexAbstractFactory {
     public Complex createComplex();
-    public Complex CreateComplex(int real, int image);
+    public Complex CreateComplex(double real, double image);
 }
